refactor(utils): move Interpolator into utility folder
